updated sample rate and shuffle times
The old version sample rate is using sample rate of the autocorrelogram to convert the spiking times to samples. I think it should be (spiking time * sample rate of ephys). Besides, I reckon shuffle one time to determine the percentile is not as reliable as 10 times shuffling.
The results seem reasonable